Extract `Hanami::API::Middleware::App` into another file
