+++ lib/hanami/api/middleware/app.rb
@@ -1,0 +1,41 @@
+# frozen_string_literal: true
+
+require "rack/builder"
+
+module Hanami
+  class API
+    module Middleware
+      # Hanami::API middleware stack
+      #
+      # @since 0.1.1
+      # @api private
+      class App
+        # @since 0.1.1
+        # @api private
+        def initialize(app, mapping)
+          @trie = Hanami::API::Middleware::Trie.new(app, "/")
+
+          mapping.each do |prefix, stack|
+            builder = Rack::Builder.new
+
+            stack.each do |middleware, args, blk|
+              builder.use(middleware, *args, &blk)
+            end
+
+            builder.run(app)
+
+            @trie.add(prefix, builder.to_app.freeze)
+          end
+
+          @trie.freeze
+        end
+
+        # @since 0.1.1
+        # @api private
+        def call(env)
+          @trie.find(env["PATH_INFO"]).call(env)
+        end
+      end
+    end
+  end
+end
